requirements/heroku.txt
-r prod.txt
Flask-SSLify==0.1.5
gunicorn==20.1.0
psycopg2-binary==2.9.3
